+++ packages/core/src/view/mixins/FoldingMixin.ts
@@ -24,6 +24,7 @@ import { toRadians } from '../../util/mathUtils';
 import Rectangle from '../geometry/Rectangle';
 import type { Graph } from '../Graph';
 import { TranslationsConfig } from '../../i18n/config';
+import { isI18nEnabled } from '../../internal/utils';
 
 type PartialGraph = Pick<
   Graph,
@@ -63,7 +64,7 @@ export const FoldingMixin: PartialType = {
     collapseToPreferredSize: true,
   },
 
-  collapseExpandResource: TranslationsConfig.isEnabled() ? 'collapse-expand' : '',
+  collapseExpandResource: isI18nEnabled() ? 'collapse-expand' : '',
 
   getCollapseExpandResource() {
     return this.collapseExpandResource;