+++ CHANGELOG.md
@@ -11,6 +11,7 @@ _**Note:** Yet to be released breaking changes appear here._
 
 **Breaking Changes**:
 - `StylesheetCodec.allowEval` is now set to `false` by default to prevent unwanted use of the eval function, as it carries a possible security risk.
+- The built-in `Translations` class is no longer used by default. To use it, call `GlobalConfig.i18n = new TranslationsAsI18n()`
 
 ## 0.16.0